refactor: rename inert.ts to inert-controller.ts
